Move to src/local-ic for go install makefile
