--- src/main.go
@@ -1,14 +1,0 @@
-package main
-
-// Used for quick script testing
-
-// go run main.go rest_server.go
-func main() {
-	// config, _ := LoadConfig("chains.json")
-	// fmt.Println(config)
-
-	// ctx := context.Background()
-	// StartNonBlockingServer(ctx, config, make(map[string]*cosmos.ChainNode))
-
-	LocalChains()
-}
